requirements_production.txt
# Core Django requirements
Django==5.2.4
openai==1.98.0
python-dotenv==1.1.1
SpeechRecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.11
psycopg2-binary==2.9.9
numpy==1.24.3
Pillow==11.3.0
gtts==2.5.1
pygame==2.5.2
scikit-learn==1.3.0

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Database and caching
redis==5.0.1
django-redis==5.4.0

# API and authentication
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1

# File storage (AWS S3)
django-storages==1.14.2
boto3==1.34.0

# Security and monitoring
django-ratelimit==4.1.0
sentry-sdk==1.40.0

# Email
django-anymail==10.2

# Environment and configuration
python-decouple==3.8
